pyproject.toml
[project]
name = "claw-observer"
version = "0.1.0"
description = "OpenClaw Gateway Observer - Sidecar + CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "websockets>=12.0",
    "pyyaml>=6.0",
    "pyjwt>=2.8.0",
    "prometheus-client>=0.19.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "paramiko>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
claw-observer = "cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sidecar", "cli"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
